refactor: 파일 분리 및 패키지 정리
