+++ app/src/main/java/com/sopt/dive/data/RepositoryProvider.kt
@@ -1,5 +1,14 @@
 package com.sopt.dive.data
 
+import com.sopt.dive.data.datasource.AuthDataSource
+import com.sopt.dive.data.datasource.UserDataSource
+import com.sopt.dive.data.datasourceimpl.AuthDataSourceImpl
+import com.sopt.dive.data.datasourceimpl.UserDataSourceImpl
+import com.sopt.dive.data.repository.AuthRepository
+import com.sopt.dive.data.repository.UserRepository
+import com.sopt.dive.data.repositoryimpl.AuthRepositoryImpl
+import com.sopt.dive.data.repositoryimpl.UserRepositoryImpl
+
 object RepositoryProvider {
     private val authDataSource: AuthDataSource by lazy {
         AuthDataSourceImpl(ServicePool.authService)
