+++ app/src/main/java/com/sopt/dive/data/ServicePool.kt
@@ -1,5 +1,8 @@
 package com.sopt.dive.data
 
+import com.sopt.dive.data.service.AuthService
+import com.sopt.dive.data.service.UserService
+
 object ServicePool {
     val authService: AuthService by lazy {
         ApiFactory.create<AuthService>()
