+++ app/src/main/java/com/sopt/dive/data/datasource/AuthDataSource.kt
@@ -1,0 +1,11 @@
+package com.sopt.dive.data.datasource
+
+import com.sopt.dive.data.BaseResponse
+import com.sopt.dive.login.LoginRequestDto
+import com.sopt.dive.login.LoginResponseDto
+
+interface AuthDataSource {
+    suspend fun postLogin(
+        request: LoginRequestDto
+    ): BaseResponse<LoginResponseDto>
+}
