+++ app/src/main/java/com/sopt/dive/data/datasourceimpl/AuthDataSourceImpl.kt
@@ -1,13 +1,10 @@
-package com.sopt.dive.data
+package com.sopt.dive.data.datasourceimpl
 
+import com.sopt.dive.data.service.AuthService
+import com.sopt.dive.data.BaseResponse
+import com.sopt.dive.data.datasource.AuthDataSource
 import com.sopt.dive.login.LoginRequestDto
 import com.sopt.dive.login.LoginResponseDto
-
-interface AuthDataSource {
-    suspend fun postLogin(
-        request: LoginRequestDto
-    ): BaseResponse<LoginResponseDto>
-}
 
 class AuthDataSourceImpl (
     private val authService: AuthService
