+++ app/src/main/java/com/sopt/dive/data/repository/AuthRepository.kt
@@ -1,0 +1,10 @@
+package com.sopt.dive.data.repository
+
+import com.sopt.dive.login.LoginModel
+import com.sopt.dive.login.LoginRequestModel
+
+interface AuthRepository {
+    suspend fun postLogin(
+        request: LoginRequestModel
+    ): Result<LoginModel>
+}
