+++ app/src/main/java/com/sopt/dive/data/repositoryimpl/UserRepositoryImpl.kt
@@ -1,5 +1,7 @@
-package com.sopt.dive.data
+package com.sopt.dive.data.repositoryimpl
 
+import com.sopt.dive.data.datasource.UserDataSource
+import com.sopt.dive.data.repository.UserRepository
 import com.sopt.dive.my.UserDataModel
 import com.sopt.dive.my.toModel
 import com.sopt.dive.signup.SignUpModel
@@ -7,16 +9,6 @@ import com.sopt.dive.signup.SignUpRequestModel
 import com.sopt.dive.signup.toDto
 import com.sopt.dive.signup.toModel
 import com.sopt.dive.util.suspendRunCatching
-
-interface UserRepository {
-    suspend fun postSignUp(
-        request: SignUpRequestModel
-    ): Result<SignUpModel>
-
-    suspend fun getUserInfo(
-        id: Long
-    ): Result<UserDataModel>
-}
 
 class UserRepositoryImpl (
     private val userDataSource: UserDataSource
