+++ app/src/main/java/com/sopt/dive/my/MyViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.sopt.dive.data.RepositoryProvider
 import com.sopt.dive.data.ServicePool
-import com.sopt.dive.data.UserDataSourceImpl
-import com.sopt.dive.data.UserRepository
-import com.sopt.dive.data.UserRepositoryImpl
+import com.sopt.dive.data.datasourceimpl.UserDataSourceImpl
+import com.sopt.dive.data.repository.UserRepository
+import com.sopt.dive.data.repositoryimpl.UserRepositoryImpl
 import com.sopt.dive.util.UiState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
